refactor: move commands into packages
