+++ internal/actions/absorb/absorb.go
@@ -1,10 +1,12 @@
-package actions
+// Package absorb provides functionality for absorbing staged changes into commits downstack.
+package absorb
 
 import (
 	"fmt"
 	"sort"
 	"strings"
 
+	"stackit.dev/stackit/internal/actions"
 	"stackit.dev/stackit/internal/engine"
 	"stackit.dev/stackit/internal/git"
 	"stackit.dev/stackit/internal/runtime"
@@ -12,16 +14,16 @@ import (
 	"stackit.dev/stackit/internal/utils"
 )
 
-// AbsorbOptions contains options for the absorb command
-type AbsorbOptions struct {
+// Options contains options for the absorb command
+type Options struct {
 	All    bool
 	DryRun bool
 	Force  bool
 	Patch  bool
 }
 
-// AbsorbAction performs the absorb operation
-func AbsorbAction(ctx *runtime.Context, opts AbsorbOptions) error {
+// Action performs the absorb operation
+func Action(ctx *runtime.Context, opts Options) error {
 	eng := ctx.Engine
 	splog := ctx.Splog
 
@@ -32,11 +34,11 @@ func AbsorbAction(ctx *runtime.Context, opts AbsorbOptions) error {
 	}
 
 	// Take snapshot before modifying the repository
-	snapshotOpts := NewSnapshot("absorb",
-		WithFlag(opts.All, "--all"),
-		WithFlag(opts.DryRun, "--dry-run"),
-		WithFlag(opts.Force, "--force"),
-		WithFlag(opts.Patch, "--patch"),
+	snapshotOpts := actions.NewSnapshot("absorb",
+		actions.WithFlag(opts.All, "--all"),
+		actions.WithFlag(opts.DryRun, "--dry-run"),
+		actions.WithFlag(opts.Force, "--force"),
+		actions.WithFlag(opts.Patch, "--patch"),
 	)
 	if err := eng.TakeSnapshot(snapshotOpts); err != nil {
 		// Log but don't fail - snapshot is best effort
@@ -251,73 +253,11 @@ func AbsorbAction(ctx *runtime.Context, opts AbsorbOptions) error {
 		upstackBranches := eng.GetRelativeStackUpstack(branch)
 
 		if len(upstackBranches) > 0 {
-			if err := RestackBranches(ctx.Context, upstackBranches, eng, splog, ctx.RepoRoot); err != nil {
+			if err := actions.RestackBranches(ctx.Context, upstackBranches, eng, splog, ctx.RepoRoot); err != nil {
 				return fmt.Errorf("failed to restack upstack branches: %w", err)
 			}
 		}
 	}
 
 	return nil
-}
-
-// printDryRunOutput prints what would be absorbed in dry-run mode
-func printDryRunOutput(hunksByCommit map[string][]git.Hunk, unabsorbedHunks []git.Hunk, eng engine.Engine, splog *tui.Splog) {
-	splog.Info("Would absorb the following changes:")
-	splog.Newline()
-
-	// Get commit info for display
-	for commitSHA, hunks := range hunksByCommit {
-		branchName, err := eng.FindBranchForCommit(commitSHA)
-		if err != nil {
-			branchName = unknown
-		}
-
-		// Get commit message - show first commit message from the branch
-		branch := eng.GetBranch(branchName)
-		commits, err := branch.GetAllCommits(engine.CommitFormatReadable)
-		if err == nil && len(commits) > 0 {
-			splog.Info("  %s in %s:", commitSHA[:8], tui.ColorBranchName(branchName, false))
-			splog.Info("    %s", commits[0])
-		} else {
-			splog.Info("  %s in %s:", commitSHA[:8], tui.ColorBranchName(branchName, false))
-		}
-
-		for _, hunk := range hunks {
-			splog.Info("    - %s (lines %d-%d)", hunk.File, hunk.NewStart, hunk.NewStart+hunk.NewCount-1)
-		}
-	}
-
-	if len(unabsorbedHunks) > 0 {
-		splog.Newline()
-		splog.Warn("The following hunks would not be absorbed:")
-		for _, hunk := range unabsorbedHunks {
-			splog.Info("  %s (lines %d-%d)", hunk.File, hunk.NewStart, hunk.NewStart+hunk.NewCount-1)
-		}
-	}
-}
-
-// printAbsorbPlan prints the plan for absorbing changes
-func printAbsorbPlan(hunksByCommit map[string][]git.Hunk, unabsorbedHunks []git.Hunk, eng engine.Engine, splog *tui.Splog) {
-	splog.Info("Will absorb the following changes:")
-	splog.Newline()
-
-	for commitSHA, hunks := range hunksByCommit {
-		branchName, err := eng.FindBranchForCommit(commitSHA)
-		if err != nil {
-			branchName = unknown
-		}
-
-		splog.Info("  Commit %s in %s:", commitSHA[:8], tui.ColorBranchName(branchName, false))
-		for _, hunk := range hunks {
-			splog.Info("    - %s (lines %d-%d)", hunk.File, hunk.NewStart, hunk.NewStart+hunk.NewCount-1)
-		}
-	}
-
-	if len(unabsorbedHunks) > 0 {
-		splog.Newline()
-		splog.Warn("The following hunks will not be absorbed:")
-		for _, hunk := range unabsorbedHunks {
-			splog.Info("  %s (lines %d-%d)", hunk.File, hunk.NewStart, hunk.NewStart+hunk.NewCount-1)
-		}
-	}
 }
