+++ internal/actions/absorb/doc.go
@@ -1,0 +1,2 @@
+// Package absorb provides functionality for absorbing staged changes into commits downstack.
+package absorb
