+++ internal/actions/create/branch_determination.go
@@ -1,0 +1,27 @@
+package create
+
+import (
+	"stackit.dev/stackit/internal/engine"
+	"stackit.dev/stackit/internal/runtime"
+	"stackit.dev/stackit/internal/utils"
+)
+
+func determineBranch(ctx *runtime.Context, opts *Options, commitMessage string, scope string) (engine.Branch, error) {
+	branchName := opts.BranchName
+	if branchName == "" {
+		// Get pattern from options (always valid, default applied in GetBranchPattern)
+		pattern := opts.BranchPattern
+
+		// Generate branch name from pattern
+		var err error
+		branchName, err = pattern.GetBranchName(ctx.Context, commitMessage, scope)
+		if err != nil {
+			return engine.Branch{}, err
+		}
+	} else {
+		// Sanitize provided branch name
+		branchName = utils.SanitizeBranchName(branchName)
+	}
+
+	return ctx.Engine.GetBranch(branchName), nil
+}
