+++ internal/actions/create/doc.go
@@ -1,0 +1,2 @@
+// Package create provides functionality for creating new stacked branches.
+package create
