+++ internal/actions/delete/doc.go
@@ -1,0 +1,2 @@
+// Package delete provides functionality for deleting branches and their metadata.
+package delete
