+++ internal/actions/doctor/doc.go
@@ -1,0 +1,2 @@
+// Package doctor provides diagnostic functionality for checking stackit environment and repository health.
+package doctor
