+++ internal/actions/fold/doc.go
@@ -1,0 +1,2 @@
+// Package fold provides functionality for folding stacked branches.
+package fold
