+++ internal/actions/fold/fold_keep.go
@@ -1,0 +1,77 @@
+package fold
+
+import (
+	"context"
+	"fmt"
+
+	"stackit.dev/stackit/internal/actions"
+	"stackit.dev/stackit/internal/engine"
+	"stackit.dev/stackit/internal/git"
+	"stackit.dev/stackit/internal/runtime"
+	"stackit.dev/stackit/internal/tui"
+)
+
+func foldWithKeep(gctx context.Context, ctx *runtime.Context, currentBranch, parent string, eng engine.Engine, splog *tui.Splog, _ Options) error {
+	// Get all children of parent (siblings + current branch)
+	parentBranch := eng.GetBranch(parent)
+	allChildren := parentBranch.GetChildren()
+
+	// Identify siblings (children of parent excluding current branch)
+	siblings := []string{}
+	for _, child := range allChildren {
+		if child.Name != currentBranch {
+			siblings = append(siblings, child.Name)
+		}
+	}
+
+	// Ensure we're on the current branch
+	currentBranchObj := eng.GetBranch(currentBranch)
+	if err := git.CheckoutBranch(gctx, currentBranchObj.Name); err != nil {
+		return fmt.Errorf("failed to checkout current branch: %w", err)
+	}
+
+	// Try fast-forward merge first, fallback to regular merge
+	_, err := git.RunGitCommandWithContext(gctx, "merge", "--ff-only", parent)
+	if err != nil {
+		// Fast-forward failed, try regular merge
+		_, err = git.RunGitCommandWithContext(gctx, "merge", "--no-edit", parent)
+		if err != nil {
+			return fmt.Errorf("failed to merge %s into %s due to conflicts. Please resolve the conflicts and run 'git commit', or abort with 'git merge --abort'", parent, currentBranch)
+		}
+	}
+
+	// Delete the parent branch (this will reparent current branch and siblings to grandparent)
+	if err := eng.DeleteBranch(gctx, parent); err != nil {
+		return fmt.Errorf("failed to delete parent branch: %w", err)
+	}
+
+	// Rebuild engine to reflect the deletion
+	if err := eng.Rebuild(eng.Trunk().Name); err != nil {
+		return fmt.Errorf("failed to rebuild engine: %w", err)
+	}
+
+	// For each sibling, set parent to current branch
+	for _, sibling := range siblings {
+		if err := eng.SetParent(gctx, sibling, currentBranch); err != nil {
+			return fmt.Errorf("failed to reparent %s to %s: %w", sibling, currentBranch, err)
+		}
+	}
+
+	splog.Info("Folded %s into %s (kept %s).",
+		tui.ColorBranchName(parent, true),
+		tui.ColorBranchName(currentBranch, false),
+		tui.ColorBranchName(currentBranch, false))
+
+	// Restack current branch and all its descendants
+	branchesToRestack := currentBranchObj.GetRelativeStack(engine.StackRange{
+		RecursiveChildren: true,
+		IncludeCurrent:    true,
+		RecursiveParents:  false,
+	})
+
+	if err := actions.RestackBranches(gctx, branchesToRestack, eng, splog, ctx.RepoRoot); err != nil {
+		return fmt.Errorf("failed to restack branches: %w", err)
+	}
+
+	return nil
+}
