+++ internal/actions/move/doc.go
@@ -1,0 +1,2 @@
+// Package move provides functionality for moving branches to different parents in the stack.
+package move
