+++ internal/actions/sync/doc.go
@@ -1,0 +1,2 @@
+// Package sync provides functionality for synchronizing stacked branches with remote repositories.
+package sync
