+++ internal/actions/undo/doc.go
@@ -1,0 +1,2 @@
+// Package undo provides functionality for undoing stackit operations using snapshots.
+package undo
